+++ example/data-persistence-tf/main.tf
@@ -11,11 +11,6 @@ variable "aws_region" {
   default = "us-east-1"
 }
 
-variable "create_service_linked_role" {
-  type    = bool
-  default = false
-}
-
 terraform {
   required_providers {
     aws = ">= 2.31.0"
@@ -31,7 +26,6 @@ module "data_persistence" {
 
   prefix                     = var.prefix
   subnet_ids                 = var.subnet_ids
-  create_service_linked_role = var.create_service_linked_role
 }
 
 output "dynamo_tables" {