+++ tf-modules/archive/variables.tf
@@ -257,14 +257,9 @@ variable "saml_launchpad_metadata_path" {
   default = "N/A"
 }
 
-variable "sts_credentials_lambda" {
-  type    = string
-  default = "gsfc-ngap-sh-s3-sts-get-keys"
-}
-
 variable "urs_url" {
   type        = string
-  default     = "https://urs.earthdata.nasa.gov/"
+  default     = "https://uat.urs.earthdata.nasa.gov/"
   description = "The URL of the Earthdata Login site"
 }
 
@@ -309,8 +304,4 @@ variable "non_complete_execution_payload_timeout" {
   description = "Number of days to retain 'non-complete' execution payload records in the database"
   type    = number
   default = 30
-}
-
-variable "region" {
-  type = string
 }