+++ tf-modules/cumulus/variables.tf
@@ -77,7 +77,8 @@ variable "prefix" {
 }
 
 variable "sts_credentials_lambda_function_arn" {
-  type = string
+  type    = string
+  default = null
 }
 
 variable "system_bucket" {
@@ -150,6 +151,18 @@ variable "custom_queues" {
   default = []
 }
 
+variable "deploy_distribution_s3_credentials_endpoint" {
+  description = "Whether or not to include the S3 credentials endpoint in the Thin Egress App"
+  type = bool
+  default = true
+}
+
+variable "distribution_api_gateway_stage" {
+  description = "The distribution API Gateway stage to create"
+  type    = string
+  default = "DEV"
+}
+
 variable "distribution_url" {
   description = " URL for the distribution API"
   type    = string
@@ -171,7 +184,7 @@ variable "ecs_cluster_instance_docker_volume_size" {
 variable "ecs_cluster_instance_image_id" {
   type        = string
   description = "AMI ID of ECS instances"
-  default     = "ami-03e7dd4efa9b91eda"
+  default     = "ami-0fa9de75aa0a1f1b3"
 }
 
 variable "ecs_cluster_instance_type" {
@@ -365,11 +378,6 @@ variable "private_archive_api_gateway" {
   default = true
 }
 
-variable "region" {
-  description = "The AWS region to deploy to"
-  type    = string
-}
-
 variable "saml_entity_id" {
   description = "The endpoint EntityID from the Launchpad Integration Request"
   type    = string
@@ -403,7 +411,7 @@ variable "throttled_queues" {
 variable "urs_url" {
   description = "The URL of the Earthdata login (URS) site"
   type        = string
-  default     = "https://urs.earthdata.nasa.gov/"
+  default     = "https://uat.urs.earthdata.nasa.gov/"
 }
 
 variable "vpc_id" {