+++ CHANGELOG.md
@@ -7,9 +7,75 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [v1.16.1] - 2019-12-6
+
+**Please note**:
+
+- The `region` argument to the `cumulus` Terraform module has been removed. You may see a warning or error if you have that variable populated.
+- Your workflow tasks should use the following versions of the CMA libraries to utilize new granule, parentArn, asyncOperationId, and stackName fields on the logs:
+  - `cumulus-message-adapter-js` version 1.0.10+
+  - `cumulus-message-adapter-python` version 1.1.1+
+  - `cumulus-message-adapter-java` version 1.2.11+
+- The `data-persistence` module no longer manages the creation of an Elasticsearch service-linked role for deploying Elasticsearch to a VPC. Follow the [deployment instructions on preparing your VPC](https://nasa.github.io/cumulus/docs/deployment/deployment-readme#vpc-subnets-and-security-group) for guidance on how to create the Elasticsearch service-linked role manually.
+- There is now a `distribution_api_gateway_stage` variable for the `tf-modules/cumulus` Terraform module that will be used as the API gateway stage name used for the distribution API (Thin Egress App)
+- Default value for the `urs_url` variable is now `https://uat.urs.earthdata.nasa.gov/` in the `tf-modules/cumulus` and `tf-modules/archive` Terraform modules. So deploying the `cumulus` module without a `urs_url` variable set will integrate your Cumulus deployment with the UAT URS environment.
+
+### Added
+
+- **CUMULUS-1563**
+  - Added `custom_domain_name` variable to `tf-modules/data-persistence` module
+- **CUMULUS-1654**
+  - Added new helpers to `@cumulus/common/execution-history`:
+    - `getStepExitedEvent()` returns the `TaskStateExited` event in a workflow execution history after the given step completion/failure event
+    - `getTaskExitedEventOutput()` returns the output message for a `TaskStateExited` event in a workflow execution history
+
+### Changed
+
+- **CUMULUS-1578**
+  - Updates SAML launchpad configuration to authorize via configured userGroup.
+   [See the NASA specific documentation (protected)](https://wiki.earthdata.nasa.gov/display/CUMULUS/Cumulus+SAML+Launchpad+Integration)
+
+- **CUMULUS-1579**
+  - Elasticsearch list queries use `match` instead of `term`. `term` had been analyzing the terms and not supporting `-` in the field values.
+
+- **CUMULUS-1619**
+  - Adds 4 new keys to `@cumulus/logger` to display granules, parentArn, asyncOperationId, and stackName.
+  - Depends on `cumulus-message-adapter-js` version 1.0.10+. Cumulus tasks updated to use this version.
+
+- **CUMULUS-1654**
+  - Changed `@cumulus/common/SfnStep.parseStepMessage()` to a static class method
+
+- **CUMULUS-1641**
+  - Added `meta.retries` and `meta.visibilityTimeout` properties to sqs-type rule. To create sqs-type rule, you're required to configure a dead-letter queue on your queue.
+  - Added `sqsMessageRemover` lambda which removes the message from SQS queue upon successful workflow execution.
+  - Updated `sqsMessageConsumer` lambda to not delete message from SQS queue, and to retry the SQS message for configured number of times.
+
+### Removed
+
+- Removed `create_service_linked_role` variable from `tf-modules/data-persistence` module.
+
+- **CUMULUS-1321**
+  - The `region` argument to the `cumulus` Terraform module has been removed
+
+### Fixed
+
+- **CUMULUS-1668** - Fixed a race condition where executions may not have been
+  added to the database correctly
+- **CUMULUS-1654** - Fixed issue with `publishReports` Lambda not including workflow execution error information for failed workflows with a single step
+- Fixed `tf-modules/cumulus` module so that the `urs_url` variable is passed on to its invocation of the `tf-modules/archive` module
+
 ## [v1.16.0] - 2019-11-15
 
 ### Added
+
+- **CUMULUS-1321**
+  - A `deploy_distribution_s3_credentials_endpoint` variable has been added to
+    the `cumulus` Terraform module. If true, the NGAP-backed S3 credentials
+    endpoint will be added to the Thin Egress App's API. Default: true
+
+- **CUMULUS-1544**
+  - Updated the `/granules/bulk` endpoint to correctly query Elasticsearch when
+  granule ids are not provided.
 
 - **CUMULUS-1580**
   - Added `/granules/bulk` endpoint to `@cumulus/api` to perform bulk actions on granules given either a list of granule ids or an Elasticsearch query and the workflow to perform.
@@ -215,6 +281,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 
 - **CUMULUS-1620** - Fixed bug where `message_adapter_version` does not correctly inject the CMA
+
+- **CUMULUS-1572** - A granule is now included in discovery results even when
+none of its files has a matching file type in the associated collection
+configuration. Previously, if all files for a granule were unmatched by a file
+type configuration, the granule was excluded from the discovery results.
+Further, added support for a `boolean` property
+`ignoreFilesConfigForDiscovery`, which controls how a granule's files are
+filtered at discovery time.
 
 ## [v1.14.2] - 2019-10-08
 
@@ -1691,7 +1765,8 @@ We may need to update the api documentation to reflect this.
 
 ## [v1.0.0] - 2018-02-23
 
-[Unreleased]: https://github.com/nasa/cumulus/compare/v1.16.0...HEAD
+[Unreleased]: https://github.com/nasa/cumulus/compare/v1.16.1...HEAD
+[v1.16.1]: https://github.com/nasa/cumulus/compare/v1.16.0...v1.16.1
 [v1.16.0]: https://github.com/nasa/cumulus/compare/v1.15.0...v1.16.0
 [v1.15.0]: https://github.com/nasa/cumulus/compare/v1.14.4...v1.15.0
 [v1.14.4]: https://github.com/nasa/cumulus/compare/v1.14.3...v1.14.4